Support multiple filestorage clients for the same file

+++ kvlayer/_file_storage.py
@@ -16,26 +16,42 @@ from kvlayer._local_memory import AbstractLocalStorage
 
 logger = logging.getLogger(__name__)
 
+
 class FileStorage(AbstractLocalStorage):
     '''
     File storage for testing and development
     '''
+
+    _datas = {}
+
     def __init__(self, *args, **kwargs):
         super(FileStorage, self).__init__(*args, **kwargs)
 
         filename = self._config['filename']
-        if self._config.get('copy_to_filename', False):
-            copy_to_filename  = self._config['copy_to_filename']
-            shutil.copyfile(filename, copy_to_filename)
-            filename = copy_to_filename
+        if filename in self._datas:
+            self._data = self._datas[filename]
+        else:
+            if self._config.get('copy_to_filename', False):
+                copy_to_filename = self._config['copy_to_filename']
+                shutil.copyfile(filename, copy_to_filename)
+                filename = copy_to_filename
 
-        if os.path.exists(filename) and os.path.getsize(filename) == 0:
-            os.remove(filename)
-        # logger.debug('Opening %s for kvlayer file storage', filename)
-        self._data = shelve.open(filename,
-                                protocol= cPickle.HIGHEST_PROTOCOL,
-                                writeback=True)
+            if os.path.exists(filename) and os.path.getsize(filename) == 0:
+                os.remove(filename)
+            self._data = shelve.open(filename,
+                                     protocol=cPickle.HIGHEST_PROTOCOL,
+                                     writeback=True)
+            self._datas[filename] = self._data
+
         self._table_names = {}
 
     def delete_namespace(self):
         self._data.clear()
+
+    def put(self, table_name, *keys_and_values, **kwargs):
+        super(FileStorage, self).put(table_name, *keys_and_values, **kwargs)
+        self._data.sync()
+
+    def delete(self, table_name, *keys):
+        super(FileStorage, self).delete(table_name, *keys)
+        self._data.sync()
